Separate HiveConnector from HdfsConnector.
