+++ modules/datacafe-server/src/main/java/edu/emory/bmi/datacafe/hdfs/HdfsConnector.java
@@ -17,7 +17,6 @@ package edu.emory.bmi.datacafe.hdfs;
 
 import edu.emory.bmi.datacafe.conf.ConfigReader;
 import edu.emory.bmi.datacafe.constants.DatacafeConstants;
-import edu.emory.bmi.datacafe.constants.HDFSConstants;
 import edu.emory.bmi.datacafe.core.CoreDataObject;
 import edu.emory.bmi.datacafe.interfaces.WarehouseConnector;
 import org.apache.logging.log4j.LogManager;
@@ -29,7 +28,6 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
-import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -94,33 +92,5 @@ public class HdfsConnector implements WarehouseConnector {
         if (ConfigReader.isRemoteDSServer()) {
             FileRemoteManager.copyFile(fileName);
         }
-    }
-
-    /**
-     * Writes to Hive
-     * @param csvFileName, file to be written to Hive
-     * @param hiveTable, the table name in Hive
-     * @param query, query to execute
-     * @throws SQLException, if execution failed.
-     */
-    public void writeToHive(String csvFileName, String hiveTable, String query)  throws SQLException {
-        try {
-            Class.forName(HDFSConstants.DRIVER_NAME);
-        } catch (ClassNotFoundException e) {
-            logger.error("Exception in finding the driver", e);
-        }
-
-        Connection con = DriverManager.getConnection(HDFSConstants.HIVE_CONNECTION_URI, HDFSConstants.HIVE_USER_NAME,
-                HDFSConstants.HIVE_PASSWORD);
-        Statement stmt = con.createStatement();
-
-        stmt.execute("drop table if exists " + hiveTable);
-
-        stmt.execute("create table " + hiveTable+ query);
-
-        String csvFilePath = HDFSConstants.HIVE_CSV_DIR + csvFileName;
-
-        String sql = "load data local inpath '" + csvFilePath + "' into table " + hiveTable;
-        stmt.execute(sql);
     }
 }
